crawler/requirements.txt
langchain_ollama>=0.0.2
langchain_elasticsearch>=0.1.0
selenium>=4.15.2
python-dotenv>=1.0.0
webdriver-manager>=4.0.1
langchain>=0.1.5
jupyterlab>=4.0.9
notebook>=7.0.6 